+++ lib/app/modules/hrm/document/screen/document_screen.dart
@@ -1,10 +1,12 @@
 import 'package:awesome_snackbar_content/awesome_snackbar_content.dart';
+import 'package:crm_flutter/app/care/constants/access_res.dart';
 import 'package:crm_flutter/app/widgets/common/indicators/crm_loading_circle.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
 import '../../../../widgets/common/dialogs/crm_delete_dialog.dart';
 import '../../../../widgets/common/messages/crm_snack_bar.dart';
+import '../../../access/controller/access_controller.dart';
 import '../controllers/document_controller.dart';
 import '../widget/document_card.dart';
 import 'add_document_screen.dart';
@@ -14,6 +16,8 @@ class DocumentScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final AccessController accessController = Get.find<AccessController>();
+
     Get.lazyPut<DocumentController>(() => DocumentController());
     final DocumentController controller = Get.find();
 
@@ -38,14 +42,17 @@ class DocumentScreen extends StatelessWidget {
 
     return Scaffold(
       appBar: AppBar(title: const Text("Documents")),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          // Navigate to add document screen
-          controller.resetForm();
-          Get.to(() => AddDocumentScreen());
-        },
-        child: const Icon(Icons.add, color: Colors.white),
-      ),
+      floatingActionButton:
+          accessController.can(AccessModule.document, AccessAction.create)
+              ? FloatingActionButton(
+                onPressed: () {
+                  // Navigate to add document screen
+                  controller.resetForm();
+                  Get.to(() => AddDocumentScreen());
+                },
+                child: const Icon(Icons.add, color: Colors.white),
+              )
+              : null,
       body: FutureBuilder(
         future: controller.loadInitial(),
         builder: (context, snapshot) {
@@ -92,32 +99,40 @@ class DocumentScreen extends StatelessWidget {
                               bottom: 8,
                               child: Row(
                                 children: [
-                                  IconButton(
-                                    icon: const Icon(
-                                      Icons.edit,
-                                      color: Colors.blue,
+                                  if (accessController.can(
+                                    AccessModule.document,
+                                    AccessAction.update,
+                                  ))
+                                    IconButton(
+                                      icon: const Icon(
+                                        Icons.edit,
+                                        color: Colors.blue,
+                                      ),
+                                      onPressed: () {
+                                        Get.to(
+                                          () => AddDocumentScreen(
+                                            document: document,
+                                            isFromEdit: true,
+                                          ),
+                                        );
+                                      },
                                     ),
-                                    onPressed: () {
-                                      Get.to(
-                                            () => AddDocumentScreen(
-                                          document: document,
-                                          isFromEdit: true,
-                                        ),
-                                      );
-                                    },
-                                  ),
-                                  IconButton(
-                                    icon: const Icon(
-                                      Icons.delete,
-                                      color: Colors.red,
+                                  if (accessController.can(
+                                    AccessModule.document,
+                                    AccessAction.delete,
+                                  ))
+                                    IconButton(
+                                      icon: const Icon(
+                                        Icons.delete,
+                                        color: Colors.red,
+                                      ),
+                                      onPressed: () {
+                                        _deleteDocument(
+                                          document.id ?? '',
+                                          document.name ?? 'Document',
+                                        );
+                                      },
                                     ),
-                                    onPressed: () {
-                                      _deleteDocument(
-                                        document.id ?? '',
-                                        document.name ?? 'Document',
-                                      );
-                                    },
-                                  ),
                                 ],
                               ),
                             ),
